src/requirements.txt
boto3==1.17.29
python-dotenv==0.15.0
PyGithub==1.54.1
